Download series and season images on sync
The series and season background image, poster and banner are
downloaded on sync so they are present when needed.

API change: the properties images, posters and banner are replaced
by the properties image, poster and banner returning the path in the
local filesystem. To get all images use get_all[images|posters|banners)